Time: 12 ms (42.48%), Space: 7.9 MB (21.68%) - LeetHub

+++ 1925-count-square-sum-triples/1925-count-square-sum-triples.cpp
@@ -14,7 +14,7 @@ public:
 
         int ans=0;
         for(double i=1;i<=n;i++){
-            for(int j=1;j<=n;j++){
+            for(double j=1;j<=n;j++){
                 double k=hypot(i,j);
                 if((int)k==k&&k>=1&&k<=n){
                     ans++;
